src/app/requirements.txt
fastapi==0.105.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
opencv